Fix code review issues: use Decimal instead of float, clarify JSON vs JSONB usage

+++ backend/schemas.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import uuid
 from datetime import datetime, date
 from typing import List, Optional, Literal
+from decimal import Decimal
 
 from pydantic import BaseModel, Field, EmailStr
 
@@ -39,8 +40,8 @@ class LocationBase(BaseModel):
     address: Optional[str] = None
     owner_info: Optional[dict] = None
     insurance_info: Optional[dict] = None
-    estimated_property_value: Optional[float] = None
-    estimated_value_with_items: Optional[float] = None
+    estimated_property_value: Optional[Decimal] = None
+    estimated_value_with_items: Optional[Decimal] = None
     location_type: Optional[str] = None
 
 
@@ -56,8 +57,8 @@ class LocationUpdate(BaseModel):
     address: Optional[str] = None
     owner_info: Optional[dict] = None
     insurance_info: Optional[dict] = None
-    estimated_property_value: Optional[float] = None
-    estimated_value_with_items: Optional[float] = None
+    estimated_property_value: Optional[Decimal] = None
+    estimated_value_with_items: Optional[Decimal] = None
     location_type: Optional[str] = None
